pyproject.toml
[project]
name = "berlin-aqi-mlops"
version = "0.1.0"
description = "Berlin AQI MLOps pipeline — predicts air quality risk categories for athletes using XGBoost, deployed on AWS."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "xgboost>=2.1.0",
    "mlflow>=2.17.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pandas>=2.2.0",
    "numpy>=2.1.0",
    "scikit-learn>=1.5.0",
    "httpx>=0.27.0",
    "boto3>=1.35.0",
    "docker>=7.1.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.56.0",
    "folium>=0.20.0",
    "streamlit-folium>=0.27.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pytest>=8.3.0",
    "ruff>=0.7.0",
]

[tool.uv]
package = false
